server/vocabularyCentral/requirements.txt
amqp==5.1.1
asgiref==3.6.0
async-timeout==4.0.2
backports.zoneinfo==0.2.1
billiard==3.6.4.0
cached-property==1.5.2
cachetools==4.2.4
celery==5.2.7
certifi==2022.12.7
chardet==3.0.4
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Deprecated==1.2.13
Django==3.2.18
django-celery-beat==2.3.0
django-celery-results==2.4.0
django-cors-headers==3.14.0
django-environ==0.4.5
django-googledrive-storage==1.6.0
django-timezone-field==5.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
dnspython==2.3.0
eventlet==0.33.1
google-api-core==2.10.2
google-api-python-client==2.86.0
google-auth==1.35.0
google-auth-httplib2==0.1.0
googleapis-common-protos==1.59.0
googletrans==4.0.0rc1
greenlet==2.0.2
gTTS==2.3.1
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2023.1.1
httpcore==0.9.1
httplib2==0.22.0
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
importlib-metadata==3.3.0
kombu==5.2.4
packaging==23.1
prompt-toolkit==3.0.38
protobuf==4.22.3
pyasn1==0.5.0
pyasn1-modules==0.3.0
PyJWT==2.6.0
pyparsing==3.0.9
python-crontab==2.7.1
python-dateutil==2.8.2
pytz==2023.3
redis==4.5.4
redis-cli==0.1.2
requests==2.28.1
rfc3986==1.5.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.4
typing_extensions==4.5.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.15
vine==5.0.0
wcwidth==0.2.6
wrapt==1.15.0
zipp==3.15.0